+++ openPER/ViewModels/DrawingKeyViewModel.cs
@@ -1,4 +1,6 @@
-﻿namespace openPER.ViewModels
+﻿using System.Security.AccessControl;
+
+namespace openPER.ViewModels
 {
     public class DrawingKeyViewModel
     {
@@ -11,9 +13,11 @@
         public int SubGroupCode { get; set; }
         public int SubSubGroupCode { get; set; }
         public int DrawingNumber { get; set; }
+        public int Revision { get; set; }
         public string ClichePartNumber { get; set; }
         public int ClichePartDrawingNumber { get; set; }
         public int ClichePartCode { get; set; }
+        public string VariantPattern { get; set; }
 
         public string FullKey => $"{ReleaseCode}/{MakeCode}/{ModelCode}/{CatalogueCode}/{GroupCode}/{SubGroupCode}/{SubSubGroupCode}/{DrawingNumber}/{ClichePartNumber}";
     }
