+++ openPER/ViewModels/PartViewModel.cs
@@ -15,7 +15,7 @@ namespace openPER.ViewModels
         public string Notes2 { get; set; }
         public string Notes3 { get; set; }
         public string Sequence { get; set; }
-        public List<ModificationModel> Modifications { get; set; }
+        public List<ModificationViewModel> Modifications { get; set; }
         public string Compatibility { get; internal set; }
         public bool IsAComponent { get; set; }
     }
