+++ openPERModels/DrawingKeyModel.cs
@@ -16,8 +16,10 @@ namespace openPERModels
         public int SubGroupCode { get; set; }
         public int SubSubGroupCode { get; set; }
         public int DrawingNumber { get; set; }
+        public int Revision { get; set; }
         public decimal ClichePartNumber { get; set; }
         public int ClichePartDrawingNumber { get; set; }
         public int ClichePartCode { get; set; }
+        public string VariantPattern { get; set; }
     }
 }
